+++ src/arritmic3d/arr3D_activations.py
@@ -1,7 +1,7 @@
 import os
 import json
 import numpy as np
-import arritmic
+import arritmic3d
 
 def load_json_list(file_path):
     """
@@ -211,6 +211,6 @@ def schedule_activation(cfg, grid, tissue):
 
     # Insert all activation times as system events
     for activation_time in activations:
-        tissue.SetSystemEvent(arritmic.SystemEventType.EXT_ACTIVATION, activation_time)
+        tissue.SetSystemEvent(arritmic3d.SystemEventType.EXT_ACTIVATION, activation_time)
 
     return activations