+++ src/arritmic3d/arritmic3D.py
@@ -8,7 +8,7 @@ import copy
 import csv
 import shutil
 
-import arritmic
+import arritmic3d
 from . import arr3D_build_slab
 
 from .arr3D_config import check_directory, get_vectorial_parameters, load_config_file, load_case_config, make_default_config, resolve_models_in_parameters
@@ -60,7 +60,7 @@ def create_tissue(grid, params):
     ncells_y = dims[1]
     ncells_z = dims[2]
 
-    tissue = arritmic.CardiacTissue(ncells_x, ncells_y, ncells_z, x_spacing, y_spacing, z_spacing)
+    tissue = arritmic3d.CardiacTissue(ncells_x, ncells_y, ncells_z, x_spacing, y_spacing, z_spacing)
 
     vparams = get_vectorial_parameters(tissue, dims, params)
     print("Parameters:", params, flush=True)
@@ -103,7 +103,7 @@ def run_simulation(case_dir, cfg):
     tissue = create_tissue(grid, cfg)
 
     # Set the timer for saving the VTK files
-    tissue.SetTimer(arritmic.SystemEventType.FILE_WRITE, cfg['VTK_OUTPUT_PERIOD'])  # time in ms
+    tissue.SetTimer(arritmic3d.SystemEventType.FILE_WRITE, cfg['VTK_OUTPUT_PERIOD'])  # time in ms
 
     # Schedule the activation protocol
     activations = schedule_activation(cfg, grid, tissue)
@@ -114,14 +114,14 @@ def run_simulation(case_dir, cfg):
         tick = tissue.update(0)
         time = tissue.GetTime()
 
-        if tick == arritmic.SystemEventType.EXT_ACTIVATION:
+        if tick == arritmic3d.SystemEventType.EXT_ACTIVATION:
             if time in activations:
                 initial_nodes = activations[time][0]
                 beat = activations[time][1]
                 tissue.ExternalActivation(initial_nodes, time, beat)
                 print("Beat at time:", time, flush=True)
 
-        elif tick == arritmic.SystemEventType.FILE_WRITE:
+        elif tick == arritmic3d.SystemEventType.FILE_WRITE:
             # Update the cell states
             grid.point_data['State'] = tissue.GetStates()
             grid.point_data['APD'] = tissue.GetAPD()