+++ test/test_multisim.py
@@ -1,4 +1,4 @@
-import arritmic
+import arritmic3d
 import numpy as np
 import pyvista as pv
 import os
@@ -6,20 +6,20 @@ import os
 # Conversion of int to CellType
 def convert_to_cell_type(cell_type):
     if cell_type == 0:
-        return arritmic.CellType.HEALTHY
+        return arritmic3d.CellType.HEALTHY
     elif cell_type == 1:
-        return arritmic.CellType.BORDER_ZONE
+        return arritmic3d.CellType.BORDER_ZONE
     else:
-        return arritmic.CellType.CORE
+        return arritmic3d.CellType.CORE
 
 # Conversion to tissue region
 def convert_to_tissue_region(region):
     if region == 0:
-        return arritmic.TissueRegion.ENDO
+        return arritmic3d.TissueRegion.ENDO
     elif region == 1:
-        return arritmic.TissueRegion.MID
+        return arritmic3d.TissueRegion.MID
     else:
-        return arritmic.TissueRegion.EPI
+        return arritmic3d.TissueRegion.EPI
 
 def main():
     vtk_file = "casos/ventricle_Tagged_2.vtk"
@@ -45,7 +45,7 @@ def main():
     ncells_x = dims[0]
     ncells_y = dims[1]
     ncells_z = dims[2]
-    tissue = arritmic.CardiacTissue(ncells_x, ncells_y, ncells_z, x_spacing, y_spacing, z_spacing)
+    tissue = arritmic3d.CardiacTissue(ncells_x, ncells_y, ncells_z, x_spacing, y_spacing, z_spacing)
 
     initial_apd = 300.0
     v_apd = [initial_apd] * (ncells_x * ncells_y * ncells_z)
@@ -72,7 +72,7 @@ def main():
     # First activation
     for i_pacing, initial_node in enumerate(p_sites_values):
         #initial_node = 12051 #tissue.GetIndex(2, 2, 1)
-        tissue = arritmic.CardiacTissue(ncells_x, ncells_y, ncells_z, x_spacing, y_spacing, z_spacing)
+        tissue = arritmic3d.CardiacTissue(ncells_x, ncells_y, ncells_z, x_spacing, y_spacing, z_spacing)
         tissue.InitPy(v_type, v_region, parameters, fiber_or)
 
         tissue.SetTimer(20)  # 20 ms